# API-CONTRACT.yaml
openapi: 3.1.0
info:
  title: Verdict API
  version: "1.0.0"
  description: >
    Canonical API contract for the Verdict hackathon judging platform.
    Every endpoint here must be implemented exactly as named — no additional
    endpoints invented by either the frontend or backend build without updating
    this file first. Enum values are defined once in ENUMS.md and referenced
    here by name. All list responses are wrapped in {"data": [...], "meta": {...}};
    all single-resource responses are wrapped in {"data": {...}}; all errors are
    {"error": {"code": "...", "message": "...", "details": {...}?}}.
servers:
  - url: /api/v1
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: verdict_session
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, enum: [unauthenticated, forbidden, not_found, validation_error, deadline_passed, event_not_open, conflict, rate_limited, internal_error] }
            message: { type: string }
            details: { type: object }
          required: [code, message]
      required: [error]
...
  parameters:
    eventSlug:
      name: event_slug
      in: path
      required: true
      schema: { type: string }
...
security:
  - sessionCookie: []
paths:
...
  /events/{event_slug}:
    parameters: [ { $ref: '#/components/parameters/eventSlug' } ]
    get:
      summary: Event detail (public)
      security: []
      responses:
        '200': { description: ok }
        '404': { description: not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    patch:
      summary: Update event settings (organizer/admin only)
      responses:
        '200': { description: ok }
        '403': { description: forbidden, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /events/{event_slug}/tracks:
    parameters: [ { $ref: '#/components/parameters/eventSlug' } ]
    get: { summary: List tracks, security: [], responses: { '200': { description: ok } } }
    post:
      summary: Create track (organizer/admin)
      responses: { '201': { description: created }, '403': { description: forbidden } }
  /tracks/{track_id}:
    parameters: [ { name: track_id, in: path, required: true, schema: { type: string, format: uuid } } ]
    patch: { summary: Update track (organizer/admin), responses: { '200': { description: ok } } }
    delete: { summary: Delete track (organizer/admin), responses: { '204': { description: no content } } }
  /events/{event_slug}/prizes:
    parameters: [ { $ref: '#/components/parameters/eventSlug' } ]
    get: { summary: List prizes, security: [], responses: { '200': { description: ok } } }
    post: { summary: Create prize (organizer/admin), responses: { '201': { description: created } } }
  /prizes/{prize_id}:
    parameters: [ { name: prize_id, in: path, required: true, schema: { type: string, format: uuid } } ]
    patch: { summary: Update prize, responses: { '200': { description: ok } } }
    delete: { summary: Delete prize, responses: { '204': { description: no content } } }

  # ---------------------------------------------------------------- TEAMS
...
  /events/{event_slug}/vote/ballot:
    parameters: [ { $ref: '#/components/parameters/eventSlug' } ]
    get:
      summary: Randomized-order project list for public voting (only while voting_open_at..voting_close_at and results not yet published)
      security: []
      responses: { '200': { description: ok }, '403': { description: voting not currently open } }
  /events/{event_slug}/vote:
    parameters: [ { $ref: '#/components/parameters/eventSlug' } ]
    post:
      summary: Cast a vote. Deduplicated at the DB level on (event_id, project_id, voter_key); rate-limited per IP and per voter_key.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                project_id: { type: string, format: uuid }
                voter_type: { type: string, enum: [email, link, account] }
                email: { type: string, description: "required if voter_type=email" }
              required: [project_id, voter_type]
      responses:
        '201': { description: created }
        '409': { description: duplicate vote, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '429': { description: rate limited, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ---------------------------------------------------------------- COMMENTS
...